requirements.txt
PyQt5
PyQtWebEngine
termcolor
qrainbowstyle==0.6.5
python-dateutil
requests
caldav==1.2.1
icalendar
recurrent
mutagen
psutil
pypiwin32; platform_system == 'Windows'
